+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.scom/vsoch/elfcall/tree/main) (0.0.x)
+ - best effort guess if an ELF binary is stripped (0.0.16)
  - tweaks to ld library path logic (0.0.15)
  - typo use_versions instead of user_versions (0.0.14)
  - add default argument to not use symbol versions (0.0.13)